FEAT: Added the jwt Token, added the token verification for door
In the jwt token based auth, the username and password verification is not yet implemented. The token refresh is not yet verified either. The API database are not yet optimized. The Unit test report generated previously will also not work due to new token based auth, hence the test report was removed from the JEnkins pipeline.