REFACTOR: Code and added user verification for the login system
Now the user is verified before new token is provided to the user. In case the email-address or the password donot match or is not present in the login_db of the server, the token will not be provided. Few endpoints were changed to formalize them. Added check for the different values so that if invalid value is entered, the appropriate exception is sent in the status response.