FEAT: Added new guest_db and car_db. Added Guest access
As per the guest access feature, added ability to add new guest. The new guest's email and passswrod will be returnrd. New end point to login using the guest access was added. Based on this a new token will be generated on guest_login. Endpoint to get all the list of guest under current user can also be accessed.